requirements.txt
django==1.7.1
psycopg2==2.5.3
selenium==2.42.1
django-pgjson==0.2.1
PyMySQL==0.6.2
django-uuidfield==0.5.0
Sphinx==1.2.3
django-nose==1.3
coverage==3.7.1
django-grappelli==2.6.3
